Don't squeeze before matching headers.
Squeeze without any arguments actually throws away ANY repeated letters, not just spaces. This is clearly not what we want. I've come to the conclusion it would be better to remove both the squeeze and the strip to avoid as much magic as possible. You can always squeeze and strip the headers externally if needed.