Pass context object to preprocessors
ATM, if one wants to access a variable with limited scope inside a
preprocessor, the only way to do that is by creating anonymous schemas
on-the-fly and take advantage from the fact that ruby blocks act as
closures. Example:

```ruby
def create_schema(offset)
  Conformist.new do
    column :numeric_value do |value|
      offset + value.to_i
    end
  end
end
```

This change allows developers to pass a context object (typically an hash, but
I'm not enforcing it) to the preprocessors, without the need to continuously
create anonymous schemas. Example:

```ruby
SCHEMA =
  Conformist.new do
    column :numeric_value do |value, context|
      context[:offset].to_i + value.to_i
    end
  end

SCHEMA.conform(something, context: {offset: 5})
```